yaim/pre_config/requirements.txt
Babel==2.6.0
pbr==4.3.0
ply==3.11
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
six==1.11.0
yaql==1.1.3
